requirements.txt
transformers
torch
python-dotenv
pyttsx3
speech_recognition
cohere